[FEAT] RedisService getCallIdByDriverId Optional 로 지정 택시기사의 상태가 변경될때마다 해당 택시기사가 가진 콜 정보를 조회합니다. 해당 택시기사가 가진 콜이 존재하지 않는 경우 NullPointerException 을 반환하여 클라이언트에게 ErrorResponse 를 응답하는 경우를 방지하게 위해 Optional 로 한번 더 처리하도록 변경하였습니다.
FIX:#53

+++ src/main/java/goorm/humandelivery/application/TaxiDriverConnectionMonitor.java
@@ -5,6 +5,7 @@ import static goorm.humandelivery.domain.model.entity.TaxiDriverStatus.*;
 import java.time.Instant;
 import java.time.ZoneId;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -71,10 +72,26 @@ public class TaxiDriverConnectionMonitor {
 				log.warn("[{}] 위치 갱신 시간 초과.", driverLoginId);
 
 				// 1. call Id 조회
-				Long callId = redisService.getCallIdByDriverId(driverLoginId);
+				Optional<String> callIdOptional = redisService.getCallIdByDriverId(driverLoginId);
+
+				if (callIdOptional.isEmpty()) {
+					// 해당 택시기사 상태 OFF_DUTY 로 DB 에서 변경..
+					TaxiType taxiType = redisService.getDriversTaxiType(driverLoginId);
+
+					// 택시기사 상태 변경에 따른 Redis 내부 처리 로직 수행
+					TaxiDriverStatus taxiDriverStatus = taxiDriverService.changeStatus(driverLoginId, OFF_DUTY);
+					redisService.handleTaxiDriverStatusInRedis(driverLoginId, taxiDriverStatus, taxiType);
+
+					log.info(
+						"[monitorReservedTaxiDrivers.TaxiDriverConnectionMonitor] 배차 실패. 택시기사 ID : {}", driverLoginId);
+					continue;
+				}
+
+
+				Long callId = Long.valueOf(callIdOptional.get());
 				String customerLoginId = callInfoService.findCustomerLoginIdById(callId);
 
-				// 2. 매칭 엔티티 삭제
+				// 매칭 엔티티 삭제
 				matchingService.deleteByCallId(callId);
 
 				// 3. 해당 택시기사 상태 OFF_DUTY 로 DB 에서 변경..
